update chap07, cote
10/1

+++ chap03/src/sec02/exam04/ArithmeticOperatorExample.java
@@ -20,7 +20,7 @@ public class ArithmeticOperatorExample {
         int result5 = v1 % v2;
         System.out.println("result5 = " + result5);
 
-        int result6 = (double) v1 / v2;
+        double result6 = (double) v1 / v2;
         System.out.println("result6 = " + result6);
     }
 }
